requirements.txt
asgiref==3.7.2
Django==4.2.8
django-admin-interface==0.27.0
django-colorfield==0.1.17
djangorestframework==3.14.0
Pillow==10.1.0
python-slugify==8.0.1
pytz==2023.3.post1
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.9.0
